requirements.txt
Django==3.0.2
django-contrib-comments==1.9.2
django-filter==2.1.0
djangorestframework==3.11.0
